+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import just_audio
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  JustAudioPlugin.register(with: registry.registrar(forPlugin: "JustAudioPlugin"))
+}