+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.5
+
+* Update dependencies and documentation.
+
+## 0.1.4
+
+* Add MacOS implementation.
+* Support cross-platform redirects on Android.
+* Bug fixes.
+
 ## 0.1.3
 
 * Fix bug in web implementation.